Setup JPA and a simple text entity
Adds a simple entity with text attributes and tests for the according
JPA repository. It is important to turn off the `@Rollback` and
`@Transactional` for the test classes that Spring would usually suggest
to rollback the database changes after the test. Otherwise, if the
transaction is never committed, the test would unfortunately never
actually hit the database and one would never see possible exceptions
(foreign keys constraints etc.) from the database.

The added test classes are already abstracted so that they can be reused
for other entities.